+++ Fanview.API/Model/TeamRanking.cs
@@ -16,7 +16,7 @@ namespace Fanview.API.Model
         public string TeamRank { get; set; }
         public string TeamName { get; set; }
         public int Kill{ get; set; }
-        public float Damage { get; set; }
+        public double Damage { get; set; }
         public int TotalPoints { get; set; }        
         public string MatchId { get; set; }
     }
